feat(app): cleaning
added back the  parameter for the on_member_update method because it threw a parameter mismatch exception. cleaned up a lot of the comments so they're more clear about what's going on. os.exit has been replaced with sys.exit cause it's safer and allows an additional message. also some of the print messages were modified slightly to be more clear about what's being done with the user.